feat: updated readme
Now lists what software is installed. Could do with each of those things being
turned into a link or having some descriptive text but w.e.

Removed 'python' from the brew installation - MacOS includes python2.x as part
of its base installation and we don't install any modules for it so it might
be a little confusing to the user.